requirements.txt
fastapi==0.128.6
python-dotenv==1.2.1
uvicorn==0.40.0
